Thrust indicator and mouse control mode
Added an indicator for the current thrust level, as well as a new control system. This system utilizes the position of the mouse to pitch/roll, and the thrust is controlled by the scroll wheel. The player camera was also modified to follow the ship more "loosely", allowing for more intuitive movement.